+++ oap-server/server-storage-plugin/storage-jdbc-hikaricp-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/jdbc/common/dao/JDBCLogQueryDAO.java
@@ -101,12 +101,17 @@ public class JDBCLogQueryDAO implements ILogQueryDAO {
                 tags.stream().map(Tag::getKey).filter(not(searchableTagKeys::contains)).collect(toSet()));
             return new Logs();
         }
+        List<String> tables;
+        if (nonNull(duration)) {
+            tables = tableHelper.getTablesForRead(
+                LogRecord.INDEX_NAME,
+                duration.getStartTimeBucket(),
+                duration.getEndTimeBucket()
+            );
+        } else {
+            tables = tableHelper.getTablesWithinTTL(LogRecord.INDEX_NAME);
+        }
 
-        final var tables = tableHelper.getTablesForRead(
-            LogRecord.INDEX_NAME,
-            duration.getStartTimeBucket(),
-            duration.getEndTimeBucket()
-        );
         final var logs = new ArrayList<Log>();
 
         for (final var table : tables) {